Inline provider dropdown + removal of save button
