+++ lib/ui/screens/home/nodeinfo_card.dart
@@ -7,21 +7,26 @@ import 'package:testapp/ui/app/app.dart';
 import 'package:testapp/ui/components/save_card.dart';
 import 'package:url_launcher/url_launcher.dart';
 import 'package:flutter/services.dart';
+
 import 'package:testapp/ui/components/inline_edit.dart';
+import 'package:testapp/ui/components/provider_dropdown.dart';
 
 class NodeInfoCard extends StatelessWidget with GetItMixin {
   NodeInfoCard({super.key});
   @override
   Widget build(BuildContext context) {
     var nodeinfo = watchOnly((NodeManagerInfo nodeinfo) => nodeinfo.nodeinfo);
-    return _MyDataTable(nodeinfo);
+    var providers = watchOnly((NodeManagerInfo info) => info.info.providers);
+
+    return _MyDataTable(nodeinfo, providers);
   }
 }
 
 class _MyDataTable extends StatefulWidget {
   final List<NodeInfo> nodeinfo;
+  final List<String> providers;
 
-  const _MyDataTable(this.nodeinfo);
+  const _MyDataTable(this.nodeinfo, this.providers);
   @override
   _MyDataTableState createState() => _MyDataTableState(nodeinfo);
 }
@@ -30,6 +35,7 @@ class _MyDataTable extends StatefulWidget {
 class _MyDataTableState extends State<_MyDataTable> {
   _MyDataTableState(this.nodeinfo);
   List<NodeInfo> nodeinfo;
+  late List<String> providers;
 
   Map<String, double> columnWidths = {
     'Name': 100,
@@ -61,7 +67,7 @@ class _MyDataTableState extends State<_MyDataTable> {
     'Status': (node) => node.status,
     // 'Added Height': (node) => node.added_height.toString(),
     // 'Confirmed Height': (node) => node.confirmed_height.toString(),
-    '': (node) => ''
+    // '': (node) => ''
   };
 
   List<NodeInfo> filteredList = [];
@@ -70,11 +76,11 @@ class _MyDataTableState extends State<_MyDataTable> {
   void initState() {
     super.initState();
     filteredList = List.from(nodeinfo);
+    providers = widget.providers;
     initNullFields();
   }
 
   void updateState() {
-    print('updateState called');
     setState(() {});
   }
 
@@ -89,7 +95,7 @@ class _MyDataTableState extends State<_MyDataTable> {
       node.name ??= 'NOT SET';
       node.provider ??= '--';
       node.price ??= -1;
-  });
+    });
   }
 
   @override
@@ -268,8 +274,22 @@ class _MyDataTableState extends State<_MyDataTable> {
     }
 
     if (e.key == 'Name') {
-      text = nameField(node, value, updateState);
+      text = inlineEdit(node, value, updateState, 'Name');
+    }
 
+    if (e.key == 'Price') {
+      text = inlineEdit(node, value, updateState, "Price");
+    }
+
+    if (e.key == 'Provider') {
+      text = ProviderDropdown(
+        providers: providers,
+        selectedProvider: value,
+        onProviderSelected: (String provider) {
+          node.provider = provider;
+          updateState();
+        },
+      );
     }
 
     return DataCell(
@@ -282,18 +302,17 @@ class _MyDataTableState extends State<_MyDataTable> {
     );
   }
 
-  dynamic nameField(node, text, reset) {
+  InlineTextEdit inlineEdit(node, text, reset, property) {
     TextEditingController _editingController =
         TextEditingController(text: text);
-    // print('in nameField: ${_editingController.text}');
-    node.name = _editingController.text;
+
     InlineTextEdit inline = InlineTextEdit(
       node: node,
       controller: _editingController,
       reset: reset,
+      property: property,
     );
     return inline;
-    // bool _isEditingText = false;
   }
 
   dynamic indexNodeMap(int index, NodeInfo node) {
