+++ lib/ui/screens/inactive/inactive_card.dart
@@ -3,14 +3,19 @@ import 'dart:io';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_base/ui/utils/bootstrap.dart';
+
 import 'package:get_it_mixin/get_it_mixin.dart';
 import 'package:testapp/api/model/inactiveInfo.dart';
 import 'package:testapp/ui/app/app.dart';
+import 'package:testapp/utils/config.dart';
+
 import 'package:http/http.dart' as http;
 import 'package:testapp/ui/components/save_card.dart';
 
-import 'package:data_table_2/data_table_2.dart';
-import 'package:testapp/utils/config.dart';
+import 'package:url_launcher/url_launcher.dart';
+import 'package:flutter/services.dart';
+import 'package:testapp/ui/components/inline_edit.dart';
+import 'package:testapp/ui/components/provider_dropdown.dart';
 import 'package:testapp/ui/screens/login/login_card.dart';
 
 class InactiveCard extends StatelessWidget with GetItMixin {
@@ -19,49 +24,90 @@ class InactiveCard extends StatelessWidget with GetItMixin {
   Widget build(BuildContext context) {
     var inactiveInfo =
         watchOnly((NodeManagerInfo inactiveInfo) => inactiveInfo.inactiveInfo);
-    return _MyDataTable(inactiveInfo);
+    var providers = watchOnly((NodeManagerInfo info) => info.info.providers);
+
+    return _MyDataTable(inactiveInfo, providers);
   }
 }
 
 class _MyDataTable extends StatefulWidget {
-  final List<InactiveInfo> nodeinfo; // Add this line to receive nodeinfo
+  final List<InactiveInfo> inactiveInfo;
+  final List<String> providers;
 
-  _MyDataTable(this.nodeinfo); // Constructor to receive nodeinfo;
+  const _MyDataTable(this.inactiveInfo, this.providers);
   @override
-  _MyDataTableState createState() => _MyDataTableState(nodeinfo);
+  _MyDataTableState createState() => _MyDataTableState(inactiveInfo);
 }
 
+// Work below //
 class _MyDataTableState extends State<_MyDataTable> {
-  _MyDataTableState(this.nodeinfo);
-  List<InactiveInfo> nodeinfo;
+  _MyDataTableState(this.inactiveInfo);
+  List<InactiveInfo> inactiveInfo;
+  late List<String> providers;
 
-  List<DataRow> allDataRows = [];
-  List<DataRow> filteredRows = [];
-  Map<String, InactiveInfo> selectedRows =
-      {}; // Holds txhash:outidx for each row
+  Map<String, double> columnWidths = {
+    'Name': 80,
+    'Provider': 80,
+    'Price': 80,
+    'IP Address': 110,
+    'Txhash': 300,
+    'Tier': 80,
+    'Status': 80,
+    'Amount': 80,
+    '': 50,
+    'Save': 50
+  };
+
+  // Holds key for each row that is selected
+  Map<String, InactiveInfo> selectedRows = {};
   DataTable? dataTable;
-  bool showButton = false; // start selected nodes button
+  bool showButton = false; // Start selected nodes button
 
   int? _sortColumnIndex;
   bool _sortAscending = true;
 
-  Color scaffoldBackgroundDark = const Color.fromARGB(255, 20, 22, 41);
-  Color cardColorDark = const Color.fromARGB(255, 58, 58, 78);
-
-  Color rowColor =
-      const Color.fromARGB(255, 14, 16, 33); // changed when selected
-
   // contains attributes to be displayed in the table
+  Map<String, String Function(InactiveInfo)> attributes = {
+    'Name': (node) => node.name ?? '',
+    'Provider': (node) => node.provider ?? '',
+    'Price': (node) => node.price.toString(),
+    'IP Address': (node) => node.ip,
+    'Txhash': (node) => node.txid,
+    'Tier': (node) => node.tier,
+    'Status': (node) => node.status,
+    'Amount': (node) => node.amount.toString(),
+    '': (node) => '',
+    // 'Save': (node) => '',
+  };
+
+  List<InactiveInfo> filteredList = [];
+
+  @override
+  void initState() {
+    super.initState();
+    filteredList = List.from(inactiveInfo);
+    providers = widget.providers;
+    initNullFields();
+  }
+
+  void updateState() {
+    setState(() {});
+  }
+
+  double initX = 0;
+  final minimumColumnWidth = 10;
+  final verticalScrollController = ScrollController();
+  final horizontalScrollController = ScrollController();
 
   void initNullFields() {
-    nodeinfo.forEach((node) => {
+    inactiveInfo.forEach((node) => {
           if (node.name == null)
             {
               node.name = 'NOT SET',
             },
           if (node.provider == null)
             {
-              node.provider = 'NOT SET',
+              node.provider = '--',
             },
           if (node.price == null)
             {
@@ -70,38 +116,14 @@ class _MyDataTableState extends State<_MyDataTable> {
         });
   }
 
-  // rebuild the table after editing nodes
-  void updateState() {
-    loadData();
-    setState(() {});
-  }
-
-  @override
-  void initState() {
-    super.initState();
-    initNullFields();
-    loadData();
-    showButton = false;
-    rowColor = scaffoldBackgroundDark;
-  }
-
-  void loadData() {
-    allDataRows = nodeinfo
-        .map<DataRow>((node) => buildDataRow(node, attributes))
-        .toList();
-    filteredRows = List.from(allDataRows);
-  }
-
   @override
   Widget build(BuildContext context) {
-    allDataRows =
-        nodeinfo.map((node) => buildDataRow(node, attributes)).toList();
     return BootstrapContainer(
       fluid: true,
       children: [
         BootstrapCol(
-          // Start all selected nodes button
-          sizes: 'col-12 col-sm-2 col-md-2 col-lg-2 col-xl-2',
+          //  White space
+          sizes: 'col-12 col-sm-6 col-md-9 col-lg-9 col-xl-9',
           child: SizedBox(
             width: 200,
             height: 50,
@@ -115,183 +137,105 @@ class _MyDataTableState extends State<_MyDataTable> {
           ),
         ),
         BootstrapCol(
-          // White space
-          child: SizedBox(),
-          sizes: 'col-12 col-sm-7 col-md-7 col-lg-7 col-xl-7',
-        ),
-        BootstrapCol(
-          // Search bar
+          // SEARCH BOX
           sizes: 'col-12 col-sm-3 col-md-3 col-lg-3 col-xl-3',
           child: Align(
             alignment: Alignment.centerRight,
             child: SizedBox(
-              width: 200,
-              child: TextField(
-                onSubmitted: filterData,
-                onChanged: (query) {
-                  if (query.isEmpty) {
-                    setState(() => filteredRows = allDataRows);
-                  }
-                },
-                decoration: const InputDecoration(
-                  labelText: "Search",
-                  suffixIcon: Icon(Icons.search),
-                ),
-              ),
-            ),
+                width: 200,
+                height: 50,
+                child: TextField(
+                  onSubmitted: filterData,
+                  onChanged: (query) {
+                    if (query.isEmpty) {
+                      setState(() => filteredList = inactiveInfo);
+                    }
+                  },
+                  decoration: const InputDecoration(
+                    labelText: "Search",
+                    suffixIcon: Icon(Icons.search),
+                  ),
+                )),
           ),
         ),
         // DATATABLE
         SizedBox(
-          height: MediaQuery.of(context).size.height * 0.6,
+          height: MediaQuery.of(context).size.height * 0.60,
           child: ConstrainedBox(
             constraints: BoxConstraints.expand(
               width: MediaQuery.of(context).size.width,
               height: MediaQuery.of(context).size.height * 0.60,
             ),
             child: SingleChildScrollView(
-              scrollDirection: Axis.vertical,
-              child: SingleChildScrollView(
-                  scrollDirection: Axis.horizontal, child: _buildDataTable()),
-            ),
+                scrollDirection: Axis.vertical, child: _buildDataTable()),
           ),
         ),
       ],
     );
   }
 
-  List<Map<String, dynamic Function(dynamic)>> attributes = [
-    {'Name': (node) => node.name},
-    {'Provider': (node) => node.provider},
-    {'Price': (node) => node.price.toString()},
-    {'IP Address': (node) => node.ip},
-    {'Txhash': (node) => node.txid},
-    {'Tier': (node) => node.tier},
-    {'Status': (node) => node.status},
-    {'Amount': (node) => node.amount.toString()},
-    // {'Vout': (node) => node.vout.toString()},
-    {'': (node) => ''},
-    {'Save': (node) => ''},
-  ];
-
-  DataTable _buildDataTable() {
-    return DataTable(
-        // decoration:
-        //     BoxDecoration(border: Border.all(color: Colors.green, width: 4)),
-        showCheckboxColumn: false,
-        showBottomBorder: true,
-        sortColumnIndex: _sortColumnIndex,
-        sortAscending: _sortAscending,
-        columnSpacing: 0,
-        columns: attributes.map((attribute) {
-          // Headerless column
-          if (attribute.keys.first == '') {
-            // return DataColumn2(size: ColumnSize.M, label: Text(''));
-            return DataColumn2(size: ColumnSize.M, label: Text(''));
-          } else if (attribute.keys.first == 'Save') {
-            return DataColumn2(size: ColumnSize.M, label: Text(''));
-          } else {
-            if (attribute.keys.first == 'Txhash') {
-              // Make the txhash column longer than the others.
-              return DataColumn(
-                label: Text(attribute.keys.first),
-              );
-            }
-            return DataColumn(
-              label: Text(attribute.keys.first),
-              onSort: (columnIndex, ascending) =>
-                  _onSort(columnIndex, ascending),
-            );
-          }
-        }).toList(),
-        rows: filteredRows);
-  }
-
   Map<String, Color> rowColors = {};
+  Color scaffoldBackgroundDark = const Color.fromARGB(255, 20, 22, 41);
+  Color cardColorDark = const Color.fromARGB(255, 58, 58, 78);
 
-  DataRow buildDataRow(node, attributes) {
-    DataCell getTextDataCell(String text, Color color) {
-      return DataCell(
-        Text(text, style: TextStyle(color: color), maxLines: 2),
-      );
-    }
-
-    var startCell = DataCell(StartNodeButton(node));
-
-    const keysToCheck = ['Name', 'Provider', 'Price'];
-    List<DataCell> cells = attributes.map<DataCell>((attribute) {
-      String featureValue = attribute.values.first(node);
-
-      // Headerless column contains startnode button
-      if (attribute.keys.first == '') {
-        return startCell;
-      } else if (attribute.keys.first == 'Save') {
-        return DataCell(SaveNodeButton(
-            node: node,
-            reset: (() {
-              updateState();
-            })));
-      } else if (keysToCheck.contains(attribute.keys.first)) {
-        if (featureValue == '-1') {
-          featureValue = 'NOT SET';
-        }
-
-        Color textColor = (featureValue == 'NOT SET' || featureValue == '--')
-            ? Colors.red
-            : Colors.green;
-        return getTextDataCell(featureValue, textColor);
-      } else if (attribute.keys.first == 'Txhash') {
-        return DataCell(
-          InkWell(
-              onTap: () async {
-                await Clipboard.setData(ClipboardData(text: node.txid));
-                ScaffoldMessenger.of(context).showSnackBar(
-                  const SnackBar(
-                    content: Text('Copied txash to clipboard'),
-                    // make this only last 1 second
-                    duration: Duration(seconds: 1),
-                  ),
-                );
-              },
-              child:
-                  Text(node.txid, style: const TextStyle(color: Colors.red))),
+  Widget _buildDataTable() {
+    return DataTable(
+      showCheckboxColumn: false,
+      showBottomBorder: true,
+      sortColumnIndex: _sortColumnIndex,
+      sortAscending: _sortAscending,
+      columnSpacing: 0,
+      dividerThickness: 0,
+      columns: attributes.entries.map((e) {
+        String title = e.key == 'Save' ? '' : e.key;
+        return DataColumn(
+          label: Stack(
+            children: [
+              Container(
+                width: columnWidths[e.key],
+                constraints: const BoxConstraints(minWidth: 10),
+                child: Text(title),
+              ),
+            ],
+          ),
+          onSort: ((columnIndex, ascending) => _onSort(columnIndex, ascending)),
         );
-      } else {
-        return DataCell(
-          Text(attribute.values.first(node),
-              style: const TextStyle(color: Colors.red)),
-        );
-      }
-    }).toList();
-
-    return DataRow(
-      color:
-          MaterialStateProperty.resolveWith<Color>((Set<MaterialState> states) {
-        return rowColors[node.txid] ?? scaffoldBackgroundDark;
-      }),
-      cells: cells,
-      selected: true,
-      onSelectChanged: (bool? selected) {
-        String myKey = node.txid;
-        if (selected != null) {
-          selected = selectedContains(node);
-          setState(() {
-            if (selected == false) {
-              rowColors[myKey] = cardColorDark;
-              selectedRows[myKey] = node;
-              if (selectedRows.length > 1) {
-                showButton = true;
-              }
-            } else {
-              rowColors[myKey] = scaffoldBackgroundDark;
-              selectedRows.remove(myKey);
-              if (selectedRows.length <= 1) {
-                showButton = false;
-              }
+      }).toList(),
+      rows: filteredList.map<DataRow>((node) {
+        return DataRow(
+          color: MaterialStateProperty.resolveWith<Color>(
+              (Set<MaterialState> states) {
+            return rowColors[node.txid] ?? scaffoldBackgroundDark;
+          }),
+          // onSelectChanged: (t) => displayPopout(node),
+          cells: attributes.entries
+              .map<DataCell>(
+                (e) => _buildCell(e, node),
+              )
+              .toList(),
+          onSelectChanged: (bool? selected) {
+            String myKey = node.txid;
+            if (selected != null) {
+              selected = selectedContains(node);
+              setState(() {
+                if (selected == false) {
+                  rowColors[myKey] = cardColorDark;
+                  selectedRows[myKey] = node;
+                  if (selectedRows.length > 1) {
+                    showButton = true;
+                  }
+                } else {
+                  rowColors[myKey] = scaffoldBackgroundDark;
+                  selectedRows.remove(myKey);
+                  if (selectedRows.length <= 1) {
+                    showButton = false;
+                  }
+                }
+              });
             }
-          });
-        }
-      },
+          },
+        );
+      }).toList(),
     );
   }
 
@@ -301,55 +245,196 @@ class _MyDataTableState extends State<_MyDataTable> {
     return contain;
   }
 
+  DataCell _buildCell(
+      MapEntry<String, String Function(InactiveInfo)> e, InactiveInfo node) {
+    // Save node button
+    if (e.key == 'Save') {
+      return DataCell(
+        SaveNodeButton(
+          node: node,
+          reset: (() {
+            updateState();
+          }),
+        ),
+      );
+    }
+
+    if (e.key == '') {
+      var startCell = DataCell(StartNodeButton(node));
+      return startCell;
+    }
+
+    String value = e.value(node);
+
+    const keysToCheck = ['Name', 'Provider', 'Price'];
+    if (keysToCheck.contains(e.key)) {
+      if (value == '-1') {
+        value = 'NOT SET';
+      }
+    }
+    Color textColor = (value == 'NOT SET' || value == '--')
+        ? Colors.red
+        : (e.key == 'IP Address')
+            ? Colors.lightBlue
+            : Colors.green;
+    Widget text = Text(
+      value,
+      style: TextStyle(color: textColor),
+      overflow: TextOverflow.ellipsis,
+      maxLines: 2,
+      softWrap: true,
+    );
+
+    if (e.key == 'IP Address' && value != 'Unknown') {
+      text = InkWell(
+        onTap: () async {
+          String port = value.contains(':') ? value.split(':')[0] : value;
+          String url = 'http://$port:16126/';
+
+          if (await canLaunchUrl(Uri.parse(url))) {
+            await launchUrl(Uri.parse(url));
+          } else {
+            throw 'Could not launch $url';
+          }
+        },
+        child: text,
+      );
+    }
+
+    if (e.key == 'Txhash') {
+      text = InkWell(
+        onTap: () async {
+          await Clipboard.setData(ClipboardData(text: node.txid));
+          showSnackBarTop(context);
+        },
+        child: text,
+      );
+    }
+
+    if (e.key == 'Name') {
+      text = inlineEdit(node, value, updateState, "Name");
+    }
+
+    if (e.key == 'Price') {
+      text = inlineEdit(node, value, updateState, "Price");
+    }
+
+    if (e.key == 'Provider') {
+      text = ProviderDropdown(
+        providers: providers,
+        selectedProvider: value,
+        onProviderSelected: (String provider) {
+          node.provider = provider;
+          updateState();
+        },
+      );
+    }
+
+    return DataCell(
+      ConstrainedBox(
+        constraints: BoxConstraints(
+          maxWidth: columnWidths[e.key]!,
+        ),
+        child: text,
+      ),
+    );
+  }
+
+  InlineTextEdit inlineEdit(node, text, reset, property) {
+    TextEditingController _editingController =
+        TextEditingController(text: text);
+
+    InlineTextEdit inline = InlineTextEdit(
+      node: node,
+      controller: _editingController,
+      reset: reset,
+      property: property,
+    );
+    return inline;
+  }
+
+  dynamic indexNodeMap(int index, InactiveInfo node) {
+    switch (index) {
+      case 0:
+        return node.name;
+      case 1:
+        return node.provider;
+      case 2:
+        return node.price;
+      case 3:
+        return node.ip;
+      case 4:
+        return node.txid;
+      case 5:
+        return node.tier;
+    }
+  }
+
+  showSnackBarTop(context) {
+    SnackBar snackBar = SnackBar(
+      content: const Text(
+        'Copied Txhash to clipboard',
+        style: TextStyle(fontSize: 20, color: Color.fromARGB(255, 12, 0, 150)),
+      ),
+      backgroundColor: Color.fromARGB(255, 204, 204, 204),
+      dismissDirection: DismissDirection.up,
+      shape: RoundedRectangleBorder(
+        borderRadius: BorderRadius.circular(10.0),
+      ),
+      behavior: SnackBarBehavior.floating,
+      margin: EdgeInsets.only(
+          bottom: MediaQuery.of(context).size.height * 0.9,
+          left: 10,
+          right: 10),
+      duration: const Duration(seconds: 1),
+    );
+
+    ScaffoldMessenger.of(context).showSnackBar(snackBar);
+  }
+
+  void _onSort(int columnIndex, bool ascending) {
+    setState(() {
+      _sortColumnIndex = columnIndex;
+      _sortAscending = ascending;
+      // Sort the data based on the selected column and sorting direction
+      filteredList.sort((a, b) {
+        final aValue = indexNodeMap(columnIndex, a);
+        final bValue = indexNodeMap(columnIndex, b);
+        return ascending
+            ? Comparable.compare(aValue, bValue)
+            : Comparable.compare(bValue, aValue);
+      });
+    });
+  }
+
+  List<dynamic> getRowData(node) {
+    return [
+      node.name,
+      node.provider,
+      node.price.toString(),
+      node.ip.toString(),
+      node.txid,
+      node.tier,
+      node.status,
+      node.amount,
+    ];
+  }
+
   void filterData(String query) {
     if (query.isEmpty) {
-      setState(() => filteredRows = allDataRows);
+      setState(() => filteredList = inactiveInfo);
     } else {
       setState(() {
-        filteredRows = allDataRows.where((row) {
-          String rowContent = row.cells.map((cell) {
-            return cell.child is Text
-                ? (cell.child as Text).data!
-                : cell.child is InkWell
-                    ? ((cell.child as InkWell).child as Text).data!
-                    : '';
-          }).join('***'); // should change this
+        // Filter based on search query
+        filteredList = inactiveInfo.where((row) {
+          String rowContent = "";
+          getRowData(row).forEach((cell) {
+            rowContent += cell.toString();
+          });
           return rowContent.toLowerCase().contains(query.toLowerCase());
         }).toList();
       });
     }
-  }
-
-  void _onSort(int columnIndex, bool ascending) {
-    setState(
-      () {
-        _sortColumnIndex = columnIndex;
-        _sortAscending = ascending;
-
-        // Sort the data based on the selected column and sorting direction
-        filteredRows.sort(
-          (a, b) {
-            final aValue = (a.cells[columnIndex].child as Text).data!;
-            final bValue = (b.cells[columnIndex].child as Text).data!;
-
-            bool isNumeric =
-                num.tryParse(aValue) != null && num.tryParse(bValue) != null;
-
-            if (isNumeric) {
-              double aNumber = double.parse(aValue);
-              double bNumber = double.parse(bValue);
-              return ascending
-                  ? aNumber.compareTo(bNumber)
-                  : bNumber.compareTo(aNumber);
-            } else {
-              return ascending
-                  ? Comparable.compare(aValue, bValue)
-                  : Comparable.compare(bValue, aValue);
-            }
-          },
-        );
-      },
-    );
   }
 }
 
@@ -361,7 +446,7 @@ class StartNodeButton extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return ElevatedButton(
-      child: const Text('Start Fluxnode'),
+      child: const Text('Start'),
       // Attempts startNode and send the api's response as popup message
       onPressed: () {
         showDialog(
